Use already existing PageNotFound component
